added task -> user relation

+++ src/Controller/TaskController.php
@@ -4,24 +4,22 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
+use App\Repository\TaskRepository;
+use App\Security\Voters\TaskVoter;
 use Doctrine\ORM\EntityManagerInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 
 class TaskController extends AbstractController
 {
-    /**
-     * @Route("/tasks", name="task_list")
-     */
-    public function list()
+    #[Route(path: '/tasks', name: 'task_list')]
+    public function list(TaskRepository $taskRepository): \Symfony\Component\HttpFoundation\Response
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository(Task::class)->findAll()]);
+        return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findAll()]);
     }
 
-    /**
-     * @Route("/tasks/create", name="task_create")
-     */
+    #[Route(path: '/tasks/create', name: 'task_create')]
     public function create(Request $request, EntityManagerInterface $em)
     {
         $task = new Task();
@@ -30,6 +28,7 @@ class TaskController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $task->setAuthor($this->getUser());
             $em->persist($task);
             $em->flush();
 
@@ -41,16 +40,19 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/tasks/{id}/edit", name="task_edit")
-     */
+    #[Route(path: '/tasks/{id}/edit', name: 'task_edit')]
     public function edit(Task $task, Request $request, EntityManagerInterface $em)
     {
+        $this->denyAccessUnlessGranted(TaskVoter::EDIT, $task);
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            if (!$task->getAuthor()) {
+                $task->setAuthor($this->getUser());
+            }
             $em->flush();
 
             $this->addFlash('success', 'La tâche a bien été modifiée.');
@@ -64,10 +66,8 @@ class TaskController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/tasks/{id}/toggle", name="task_toggle")
-     */
-    public function toggleTask(Task $task, EntityManagerInterface $em)
+    #[Route(path: '/tasks/{id}/toggle', name: 'task_toggle')]
+    public function toggleTask(Task $task, EntityManagerInterface $em): \Symfony\Component\HttpFoundation\RedirectResponse
     {
         $task->toggle(!$task->isDone());
         $em->flush();
@@ -77,11 +77,11 @@ class TaskController extends AbstractController
         return $this->redirectToRoute('task_list');
     }
 
-    /**
-     * @Route("/tasks/{id}/delete", name="task_delete")
-     */
-    public function deleteTask(Task $task, EntityManagerInterface $em)
+    #[Route(path: '/tasks/{id}/delete', name: 'task_delete')]
+    public function deleteTask(Task $task, EntityManagerInterface $em): \Symfony\Component\HttpFoundation\RedirectResponse
     {
+        $this->denyAccessUnlessGranted(TaskVoter::DELETE, $task);
+
         $em->remove($task);
         $em->flush();
